try to wrap `PageIndex` in option as much as possible
We have asserts everywhere for `page.is_valid()`. It would be preferable
to have use `Option<T>` to represent valid vs invalid pages. However, I
don't have a good solution for serialization of options without
requiring an extra byte. We maintain the hack for the meta_page stuff
for now.