split `Node` trait into `Node` with a GAT for `NodeView<'a>` to help with lifetime issues
In the previous trait if we wanted a method such as
`fn f(&mut self, other: &mut Self)`, `Self` would refer to `Node<'a, K, V>` with a fixed lifetime `'a`.
However, we don't want to require these two lifetimes to be the same.

Now we can write `fn f(this: &mut Self::View<'_>, other: &mut Self::View<'_>')` which gives distinct lifetimes to each parameter avoiding the unnecessary coupling.
I'm not aware of a better solution to this than using GATs currently.