random refactor of btree?
This started off as trying to remove the `Clone` bounds from `K, V` but
not sure where it ended up...

The insight inspired by looking at rkyv source code was to define the
`Archive` implementation for `KeyValuePair` manually in a way that only
required references to `K` and `V`. This allows us to avoid the clones.

`impl<'a, K: Archive, V: Archive> Archive for KeyValuePair<&'a K, &'a V> `

As a trade off, we now have to deserialize the key sometimes which seems
like an even worse trade off.

There is no significant difference in benchmarks so at least.

Also, we introduced `'static` bounds everywhere on `K, V`. I don't think
this is necessary but not sure how to make the compiler happy about
without it.