[btree] implement deletion by simply removing the slot
The previous implementation was misguided as we were trying to preserve
slot indices similar to how heap slotted pages must behave. But this
requirement doesn't exist for sorted btree slotted pages so we just
remove it instead.

The previous implementation was incomplete as splitting a node would
effectively undelete any entries marked as removed as it forgot to check
the flag before copying slots over.

The new approach also works well with the existing defragmentation impl
and nothing needs to change there. Any lost space will be reclaimed by it.