add in locally inserted when scanning table
Tests are mostly passing but it's still completely broken as the tests don't catch the issues.

One issue is that transactions won't be able to see their own version of a tuple that is updated by another transaction
as it gets overwritten in place with a newer tuple with no way to retrieve the overwritten tuple.

We should maybe relax the versioning and implement read committed for now?