give up on backends that require mutable references to transactions
This includes `heed` currently, although not sure why it actually needs
a mutable reference. We can roll with redb for the short term.

It is too difficult to parameterize over & and &mut. It's bad enough
with all the other stuff we have to abstract over and this one is not
tenable