add `dedup cte` optimization pass
After shuffling plans all around the place we can end up duplicating cte nodes.
We add a pass to only keep the first cte node for each name (which is required to be unique anyway).

There is probably a better solution where we represent ctes differently such that this can't occur but this is an easy solution for now.