dearcify all the physical nodes and use a box instead
add non-mutable conversion methods for PhysicalNode and a macro to implement them

All tests passing.

One of the primary points of this whole exercise was to allow mutable access
for the operator/source/sink methods but there are still some problems to solve there.

However, the other reason to do this was to have an id per node which is useful for storing metadata in hashmaps for example.